Rename CommonDefs to Common and clean up unused headers
