+++ AnalysisTools/DefaultAnalysis_tool.cc
@@ -12,15 +12,15 @@
 #include "nusimdata/SimulationBase/MCFlux.h"
 #include "larcore/Geometry/Geometry.h"
 #include "larsim/EventWeight/Base/MCEventWeight.h"
-#include "../CommonDefs/BacktrackingFuncs.h"
-#include "../CommonDefs/Geometry.h"
-#include "../CommonDefs/SCECorrections.h"
-#include "../CommonDefs/Containment.h"
-#include "../CommonDefs/TrackShowerScoreFuncs.h"
-#include "../CommonDefs/ProximityClustering.h"
-#include "../CommonDefs/Descendents.h"
-#include "../CommonDefs/Scatters.h"
-#include "../CommonDefs/Pandora.h"
+#include "../Common/BacktrackingFuncs.h"
+#include "../Common/Geometry.h"
+#include "../Common/SCECorrections.h"
+#include "../Common/Containment.h"
+#include "../Common/TrackShowerScoreFuncs.h"
+#include "../Common/ProximityClustering.h"
+#include "../Common/Descendents.h"
+#include "../Common/Scatters.h"
+#include "../Common/Pandora.h"
 #include "ubobj/Optical/UbooneOpticalFilter.h"
 #include "canvas/Persistency/Common/TriggerResults.h"
 #include "larpandora/LArPandoraInterface/LArPandoraHelper.h"
