+++ CMakeLists.txt
@@ -16,7 +16,7 @@ include_directories(/cvmfs/larsoft.opensciencegrid.org/products/python/v2_7_14b/
 set(PYTHON_LIB_DIR /cvmfs/larsoft.opensciencegrid.org/products/python/v2_7_14b/Linux64bit+3.10-2.17/lib)
 set(PYTHON_LIBRARY ${PYTHON_LIB_DIR}/libpython2.7.so)
 
-add_subdirectory(CommonDefs)
+add_subdirectory(Common)
 add_subdirectory(SelectionTools)
 add_subdirectory(AnalysisTools)
