+++ Common/BacktrackingFuncs.h
