+++ Common/TrackShowerScoreFuncs.h
