--- CommonDefs/LLRPID_electron_photon_lookup.h
@@ -1,184 +1,0 @@
-#ifndef ELECTRON_PHOTON_LOOKUP_H
-#define ELECTRON_PHOTON_LOOKUP_H
-#include <stdlib.h>
-#include <vector>
-
-namespace common
-{
-  struct ElectronPhotonLookUpParameters
-  {
-    std::vector<float> dedx_edges_pl_0 = {
-    0.000, 0.316, 0.632, 0.947, 1.263, 1.579, 1.895, 2.211, 2.526, 2.842, 3.158, 3.474, 3.789, 4.105, 4.421, 4.737, 5.053, 5.368, 5.684, 6.000,
-    6.500, 7.000, 7.500, 8.000, 9.000, 10.000, 12.000, 15.000, 20.000, 30.000, 50.000,
-    };
-
-
-    std::vector<std::vector<float>> parameters_edges_pl_0 = {
-    {0.000, 1.000, 2.000, 3.000, 4.000, 5.000, 10.000, 1000.000, },
-    {0.300, 0.600, 1.000, 1.500, 3000.000, },
-    };
-
-    std::vector<float> dedx_pdf_pl_0 = {
-    0.677, -0.110, -0.126, -0.068, 0.685, 1.159, 1.103, 0.725, -0.111, -0.434, -1.373, -1.500, -1.651, -1.182, -1.103, -1.150, -0.683, -0.205, -0.535, -0.285,
-    0.215, 0.409, 0.294, 0.804, 1.556, 1.301, 1.194, 2.329, 2.889, 2.014, 0.005, -0.326, -0.058, 0.274, 0.183, 0.704, 0.759, -0.082, -0.332, -0.339,
-    -0.871, -0.776, -1.477, -0.643, -0.406, -0.152, 0.052, -0.401, 0.148, 0.486, 0.816, 0.515, 1.326, 1.933, 1.391, 2.530, 0.000, 0.000, 0.921, 0.000,
-    -0.529, -0.165, 0.410, -0.219, -0.059, 0.625, 0.025, -0.529, 0.059, -0.445, -0.423, -0.018, -0.386, 0.793, 0.675, 1.081, 0.000, 0.675, 0.000, 0.000,
-    1.774, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.053, 0.176, -0.727, 0.363, -0.517, -0.533, 0.026, -0.580, 0.432, 0.000,
-    1.125, 1.125, 1.125, 0.000, 0.000, 1.818, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    0.000, -0.126, -0.189, 0.300, 0.949, 1.387, 1.484, 1.118, 0.453, -0.228, -0.956, -1.506, -1.527, -1.856, -1.569, -1.351, -1.400, -1.543, -1.380, -1.155,
-    -0.731, -0.584, -0.510, -0.251, -0.554, 0.818, 0.841, 0.823, 1.753, 0.000, 0.203, -0.397, -0.322, -0.128, 0.543, 1.208, 1.484, 0.734, 0.083, -0.580,
-    -0.833, -1.781, -1.475, -1.429, -1.001, -1.064, -0.844, -0.790, -0.949, -1.013, -0.159, -0.084, 0.049, 0.159, 0.609, 1.233, 0.203, 2.842, 0.897, 0.000,
-    0.070, 0.109, 0.435, 0.573, 0.935, 0.617, 1.421, -0.163, -0.243, -0.952, -1.193, -0.910, -0.674, -0.451, -1.262, -1.144, -1.462, -0.769, 0.000, -0.451,
-    0.242, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, -0.201, 0.474, -0.124, 0.524, -0.913, -0.019, -0.594, 0.610, 0.792, -0.307,
-    0.000, -1.154, 0.792, 0.099, 0.000, 0.099, 0.000, 0.000, 0.000, 0.792, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    2.458, -0.342, -0.234, 0.051, 0.541, 1.012, 1.082, 0.897, 0.416, -0.213, -0.710, -1.018, -1.631, -1.612, -1.488, -1.211, -1.254, -1.136, -1.224, -0.583,
-    -0.831, -0.684, -1.137, -0.808, -0.300, 0.072, 0.524, 0.444, 1.765, 2.864, 0.732, -0.759, -0.593, -0.015, 0.456, 1.063, 1.298, 0.609, 0.799, -0.517,
-    -0.638, -1.131, -1.439, -1.247, -2.175, -1.214, -2.719, -0.799, -0.969, -0.703, -0.942, -1.187, -0.759, 0.203, 0.262, 0.850, -0.249, 0.850, 0.850, 0.000,
-    -0.831, -0.160, -0.169, 0.150, 0.725, 0.876, 1.302, 0.500, 0.463, 0.224, -0.677, -1.054, -1.297, -0.990, -2.489, -2.293, -2.102, -1.555, 0.000, -0.661,
-    -0.661, -0.543, -0.543, -0.073, 0.843, 0.000, 0.000, 0.000, 0.000, 0.000, -0.124, -1.050, -0.706, 0.423, 0.859, 0.821, 0.741, 0.924, -0.357, -0.270,
-    -0.175, 0.000, -1.205, 0.048, 0.741, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    0.000, -0.568, -0.622, -0.341, 0.216, 0.787, 0.766, 0.684, 0.108, -0.026, -0.560, -0.937, -1.456, -1.253, -1.068, -0.946, -0.722, -0.547, -0.294, -0.540,
-    -0.269, -1.019, -0.758, -0.397, -0.781, -0.343, -0.292, -0.617, 0.664, 0.000, -0.655, -0.480, -0.441, -0.945, 0.222, 0.717, 0.876, 0.431, 0.257, -0.132,
-    -0.389, -0.815, -1.124, -0.764, -0.850, -0.506, -0.596, -0.600, -1.049, -0.600, -0.923, -0.230, 0.493, -0.250, -0.575, 0.118, 0.811, 0.811, 0.811, 0.000,
-    -0.564, -0.935, -0.139, -0.242, 0.161, 0.622, 1.293, 1.100, 0.292, -0.219, 0.187, -0.836, 0.000, -1.620, -1.429, 0.000, 0.000, -0.094, -0.681, -0.969,
-    -0.787, -1.257, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.007, 0.517, 0.007, -0.642, 0.959, 0.051, 1.211, 0.294, -0.176, -0.581,
-    0.112, -0.294, 0.000, 0.000, 0.000, 0.000, -0.294, 0.805, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    -0.058, -0.400, -0.820, -0.896, 0.086, 0.392, 0.658, 0.447, 0.229, 0.124, -0.624, -0.886, -0.815, -0.774, -0.858, -1.064, -0.696, -0.058, -0.201, -0.015,
-    -0.195, -0.070, -0.101, -0.535, -0.333, -0.444, 0.252, -0.406, -0.240, 2.139, 0.294, -0.737, -0.847, -1.158, -0.009, 0.581, 0.839, 0.471, 0.122, 0.086,
-    -0.520, -0.421, -0.820, -1.373, -0.668, -1.604, -1.118, -0.294, -0.550, -0.520, 0.041, 0.173, 0.173, 0.491, 0.086, 0.528, -0.474, 0.086, 1.878, 0.000,
-    -0.541, -0.928, -0.307, -0.960, 0.024, 0.843, 1.063, 0.678, -0.264, -0.129, -0.158, 0.011, -1.034, -1.545, -0.421, -0.198, 0.000, -1.545, -0.541, -0.746,
-    0.000, 1.046, 0.000, 1.451, -0.341, 0.000, 0.000, 0.000, 0.000, 0.000, -0.152, -0.321, 0.250, 0.224, 0.735, 0.196, 1.988, 0.534, 0.147, 0.330,
-    -0.651, -0.112, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    -0.625, -1.070, -0.714, -0.655, -0.218, 0.081, 0.200, 0.069, -0.038, -0.150, -0.226, -0.258, -0.378, -0.436, -0.166, -0.102, -0.132, 0.103, 0.303, 0.519,
-    0.496, 0.395, 0.330, 0.334, 0.576, 0.366, 0.610, 0.803, 0.890, 0.597, -0.334, -1.221, -1.103, -1.181, -0.481, 0.203, 0.365, 0.193, -0.041, -0.031,
-    -0.147, -0.004, -0.044, -0.321, 0.008, 0.141, 0.159, 0.446, 0.285, 0.510, 0.160, 0.590, 0.818, 0.624, 0.495, 0.468, 0.901, 1.458, 0.000, 1.389,
-    -0.523, -0.905, -0.808, -0.829, -0.649, 0.404, 0.694, 0.737, 0.443, 0.289, 0.104, -0.286, 0.226, -0.492, -0.341, -0.392, -0.426, -0.789, 0.118, 0.044,
-    0.177, 0.642, 0.583, 0.044, 0.149, 0.449, -0.649, 0.000, 0.000, 0.000, -0.246, -0.815, -0.325, -0.154, -0.135, 0.203, 0.586, 0.689, 0.286, 0.385,
-    0.586, 0.787, 0.298, 1.038, -0.618, 0.298, -0.330, 0.586, 0.404, 0.809, 0.586, 0.586, 0.000, 0.000, 0.000, 0.586, 0.000, 0.000, 0.000, 0.000,
-    0.180, 0.062, -0.122, -0.244, -0.438, -0.474, -0.423, -0.378, -0.295, -0.156, -0.074, 0.009, 0.025, 0.119, 0.210, 0.309, 0.373, 0.419, 0.539, 0.652,
-    0.734, 0.767, 0.827, 1.052, 1.189, 1.343, 1.647, 2.184, 2.563, 1.935, -0.236, -0.358, -0.568, -0.595, -0.609, -0.388, -0.304, -0.361, -0.173, -0.128,
-    -0.073, 0.091, 0.214, 0.314, 0.430, 0.486, 0.506, 0.672, 0.848, 1.055, 0.977, 1.103, 1.270, 1.627, 1.784, 1.732, 2.309, 2.640, 3.605, 1.959,
-    -0.378, -0.541, -0.467, -0.456, -0.393, -0.100, 0.084, 0.016, -0.022, 0.119, 0.175, 0.470, 0.592, 0.653, 0.982, 0.835, 0.972, 0.767, 0.935, 1.303,
-    1.101, 1.314, 1.843, 1.134, 1.479, 1.458, 1.801, 2.505, 0.000, 0.000, -0.259, -0.449, -0.322, -0.360, -0.139, -0.007, 0.395, 0.563, 0.496, 0.842,
-    0.860, 0.759, 1.146, 1.137, 1.471, 1.801, 1.062, 1.618, 1.676, 1.270, 1.137, 1.830, 1.424, 3.216, 0.000, 2.326, 1.493, 0.000, 0.000, 0.000,
-    };
-
-
-    std::vector<float> dedx_edges_pl_1 = {
-    0.000, 0.316, 0.632, 0.947, 1.263, 1.579, 1.895, 2.211, 2.526, 2.842, 3.158, 3.474, 3.789, 4.105, 4.421, 4.737, 5.053, 5.368, 5.684, 6.000,
-    6.500, 7.000, 7.500, 8.000, 9.000, 10.000, 12.000, 15.000, 20.000, 30.000, 50.000,
-    };
-
-
-    std::vector<std::vector<float>> parameters_edges_pl_1 = {
-    {0.000, 1.000, 2.000, 3.000, 4.000, 5.000, 10.000, 1000.000, },
-    {0.300, 0.600, 1.000, 1.500, 3000.000, },
-    };
-
-    std::vector<float> dedx_pdf_pl_1 = {
-    1.056, -0.216, -0.311, 0.261, 0.569, 1.007, 1.006, 0.795, 0.222, -0.481, -1.050, -1.418, -1.259, -1.438, -1.406, -0.780, -1.049, -0.944, -0.370, -0.565,
-    0.013, 0.249, 0.544, 0.637, 0.906, 1.338, 1.780, 2.726, 3.134, 3.166, -0.577, -0.285, 0.048, 0.303, 0.727, 0.865, 0.931, 0.004, -0.525, -0.591,
-    -1.125, -0.974, -1.125, -1.001, -0.938, -0.821, -0.188, -0.075, -0.144, -0.015, 0.200, 0.787, 0.366, 1.647, 0.954, 1.647, 1.935, 0.954, 2.053, 0.000,
-    -0.205, -0.539, -0.119, 0.003, 0.033, 0.367, 0.494, 0.116, -0.385, -0.172, -0.808, -0.172, 0.308, -0.459, 1.215, 0.416, -0.172, 0.234, 0.000, 0.234,
-    0.000, 0.927, 0.927, 0.927, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, -0.257, 0.116, -0.352, -0.188, 0.684, 0.187, -1.069, -0.152, 0.358, -0.740,
-    1.745, 2.661, 1.052, 0.000, 0.358, 0.000, 0.000, 0.000, 0.000, 0.000, 1.052, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    1.779, 0.275, 0.281, 0.297, 0.933, 1.293, 1.336, 1.102, 0.561, -0.233, -0.966, -1.638, -1.722, -1.813, -1.390, -1.157, -1.489, -1.561, -0.819, -0.805,
-    -0.647, -0.607, -0.177, -0.167, 0.271, 0.616, 0.734, 1.625, 2.734, 0.000, -1.607, -0.383, 0.137, 0.402, 0.920, 1.369, 1.266, 0.958, 0.333, -0.336,
-    -0.961, -2.055, -1.671, -1.703, -1.476, -1.769, -0.821, -2.350, -1.361, -0.956, -0.397, -0.806, 0.110, -0.266, -1.345, 0.804, 0.735, 0.958, -0.141, 0.000,
-    -0.641, -0.162, 0.480, -0.357, 1.191, 1.027, 0.898, 0.303, -0.020, -0.450, -1.740, -0.406, -1.335, -1.174, -1.335, -1.845, 0.000, -1.622, 0.170, 0.170,
-    0.170, -0.236, 0.000, 0.000, 0.000, 0.170, 0.000, 0.000, 0.000, 0.000, 0.192, 0.102, -0.077, -0.262, 0.440, 0.607, -0.059, 0.010, 0.192, 0.010,
-    0.000, -1.377, 0.000, 0.703, 0.000, -0.396, 0.000, 0.000, 0.703, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    -0.321, -0.901, -0.356, 0.183, 0.595, 0.854, 1.000, 0.892, 0.432, -0.120, -0.632, -1.244, -1.607, -1.277, -1.507, -1.094, -0.939, -0.992, -1.168, -0.732,
-    -0.693, -0.920, -0.060, -0.459, -0.202, -0.309, 0.554, 0.737, 1.640, 1.353, -2.114, -1.321, -0.403, 0.107, 0.649, 1.378, 1.069, 0.814, 0.413, -0.256,
-    -0.841, -0.990, -1.870, -1.049, -2.268, -1.632, -1.121, -2.315, -1.303, -1.188, -1.016, -0.610, -1.863, -0.536, -0.764, -0.051, 0.000, 0.083, 0.000, 0.000,
-    -0.771, -0.009, 0.061, 0.342, 0.678, 0.874, 1.003, 0.879, 0.359, -0.260, -0.953, -1.154, -1.282, -2.157, 0.000, -2.052, -1.107, -1.464, 0.000, -0.771,
-    -0.548, 0.000, 0.000, 0.433, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.148, -0.351, 0.126, 0.059, 0.618, -0.337, 1.157, 0.841, -0.298, -0.768,
-    0.618, 0.213, -0.768, 0.000, 0.000, 0.000, -1.174, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, -0.075, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    0.326, -0.030, -0.759, -0.392, 0.319, 0.666, 0.605, 0.643, 0.513, -0.036, -0.545, -0.774, -1.058, -1.081, -1.573, -1.198, -0.699, -0.656, -0.521, -0.327,
-    -0.417, -0.698, -0.202, -0.501, -0.189, 0.213, 0.509, 0.509, 1.936, 2.118, -2.087, -0.807, -0.858, -0.365, 0.344, 0.891, 1.084, 0.712, 0.048, -0.156,
-    -0.916, -1.069, -0.880, -1.124, -1.519, -1.212, -1.212, -0.560, -0.965, -0.632, -0.989, -1.069, 0.317, -0.360, -1.212, -0.401, 0.110, 0.000, 0.000, 0.000,
-    0.168, -0.391, -0.516, -0.642, -0.087, 0.420, 1.182, 0.791, 1.070, 0.120, -1.164, -0.679, -0.720, -0.679, -2.001, -0.168, -2.001, -0.902, -0.679, -1.544,
-    0.000, 0.000, 0.000, -0.209, 0.014, -0.391, 0.000, 0.000, 0.000, 0.000, -0.372, -0.155, 0.034, 0.167, 0.727, 0.727, 0.881, 0.409, -0.084, -0.189,
-    0.034, -1.576, 0.000, 0.000, -0.882, 0.727, 0.000, 0.727, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    0.299, -1.205, -1.075, -0.598, 0.073, 0.438, 0.474, 0.441, 0.269, 0.129, -0.290, -0.863, -0.714, -0.737, -0.725, -0.700, -0.903, -0.528, -0.444, -0.136,
-    -0.176, 0.029, -0.813, -0.420, 0.468, 0.029, 0.299, 0.741, 0.992, 0.000, -1.029, -1.012, -1.403, -0.480, -0.067, 0.558, 0.768, 0.621, 0.286, -0.300,
-    -0.603, -0.533, -0.474, -1.141, -0.950, -1.029, -0.472, -0.423, 0.237, 0.170, -0.654, -0.846, -0.175, -0.536, -1.316, 0.070, 0.252, 1.456, 0.000, 0.000,
-    0.043, -0.650, -0.981, -0.552, 0.202, 0.526, 0.736, 0.536, 0.235, -0.219, -0.427, -0.427, -0.922, -0.586, -0.180, 0.043, 0.177, 0.000, -1.749, -0.468,
-    -0.650, 0.331, 0.043, 0.000, -0.650, 0.000, -0.362, 0.000, 0.000, 0.000, -0.423, 0.214, -0.215, -1.511, 0.645, 0.517, 1.028, 0.404, 0.134, 1.251,
-    0.229, -0.176, -0.358, 0.047, -0.869, 0.740, 0.000, 0.000, 0.000, 0.000, 0.740, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    -0.728, -1.120, -0.986, -0.567, -0.295, 0.050, 0.139, 0.202, 0.088, 0.034, -0.150, -0.284, -0.388, -0.326, -0.275, -0.196, -0.011, 0.117, 0.329, 0.343,
-    0.354, 0.313, 0.204, 0.320, 0.481, 0.743, 0.687, 0.803, 1.117, 1.932, -1.361, -1.327, -1.129, -1.107, -0.446, 0.208, 0.311, 0.284, 0.082, -0.012,
-    -0.136, -0.258, -0.307, -0.233, -0.119, -0.193, 0.270, 0.333, 0.256, 0.441, 0.549, 0.627, 0.645, 0.813, 0.332, 1.085, 0.698, 0.649, 1.391, 0.698,
-    -0.868, -0.989, -1.059, -0.688, -0.260, 0.353, 0.558, 0.555, 0.165, 0.278, -0.185, 0.189, 0.001, -0.331, -0.120, 0.089, -0.218, 0.453, 0.620, -0.168,
-    -0.027, 0.263, 0.821, -0.027, -0.073, 0.001, 0.802, 0.000, 0.000, 0.000, -0.226, -0.450, -0.317, -0.641, -0.346, 0.302, 1.098, 0.635, 0.312, 0.517,
-    0.635, -0.555, -0.252, 0.229, -0.618, -0.138, 1.734, -0.975, 0.000, -0.176, -0.058, 1.040, 2.021, 0.000, 0.635, 0.229, 0.000, 0.000, 0.000, 0.000,
-    0.178, 0.031, -0.178, -0.320, -0.455, -0.438, -0.382, -0.367, -0.247, -0.168, -0.071, 0.028, 0.092, 0.123, 0.206, 0.243, 0.361, 0.417, 0.580, 0.559,
-    0.710, 0.734, 0.843, 1.073, 1.163, 1.435, 1.750, 2.087, 1.918, 1.146, -0.307, -0.435, -0.420, -0.577, -0.553, -0.430, -0.292, -0.249, -0.282, -0.102,
-    -0.048, 0.009, 0.124, 0.230, 0.247, 0.404, 0.515, 0.456, 0.643, 0.817, 0.995, 1.001, 1.212, 1.237, 1.614, 1.565, 2.207, 2.807, 3.398, 2.271,
-    -0.523, -0.592, -0.607, -0.469, -0.370, -0.147, 0.000, 0.065, 0.148, 0.188, 0.126, 0.590, 0.650, 0.594, 0.680, 0.673, 0.667, 1.073, 0.833, 0.985,
-    1.221, 1.197, 1.354, 1.509, 1.307, 1.968, 2.498, 2.038, 0.000, 0.000, -0.253, -0.450, -0.385, -0.203, -0.221, 0.189, 0.427, 0.313, 0.544, 0.776,
-    0.747, 0.440, 0.481, 0.893, 1.064, 1.921, 0.403, 1.732, 1.287, 2.211, 1.861, 2.534, 1.508, 1.469, 1.141, 1.907, 0.000, 0.000, 0.000, 0.000,
-    };
-
-
-    std::vector<float> dedx_edges_pl_2 = {
-    0.000, 0.316, 0.632, 0.947, 1.263, 1.579, 1.895, 2.211, 2.526, 2.842, 3.158, 3.474, 3.789, 4.105, 4.421, 4.737, 5.053, 5.368, 5.684, 6.000,
-    6.500, 7.000, 7.500, 8.000, 9.000, 10.000, 12.000, 15.000, 20.000, 30.000, 50.000,
-    };
-
-
-    std::vector<std::vector<float>> parameters_edges_pl_2 = {
-    {0.000, 1.000, 2.000, 3.000, 4.000, 5.000, 10.000, 1000.000, },
-    {0.300, 0.600, 1.000, 1.500, 3000.000, },
-    };
-
-    std::vector<float> dedx_pdf_pl_2 = {
-    1.329, -0.542, -1.327, -0.981, -0.253, 1.257, 1.643, 0.809, 0.083, -0.358, -0.940, -1.776, -2.043, -1.676, -1.270, -1.084, -0.822, -0.651, -0.580, -0.171,
-    0.193, 0.216, 0.449, 0.955, 1.692, 1.696, 2.088, 2.946, 3.057, 4.507, 0.000, -0.693, -1.692, -0.259, -0.269, 1.073, 1.523, 0.929, -0.018, -1.086,
-    -1.521, -1.492, -1.646, -1.077, -1.869, -1.326, -0.878, -0.928, -0.839, -0.582, -0.048, 0.240, -0.393, 0.706, 1.687, 0.000, 1.974, 0.000, 0.000, 0.000,
-    0.000, -1.532, -0.382, -0.721, -0.300, 0.588, 1.484, 0.680, 0.221, -0.696, -1.350, -1.021, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 2.562, 0.000,
-    0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, -0.377, 0.075, -0.022, 0.468, -0.198, 0.635, -0.407, -0.058, -0.058, 0.229,
-    0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    0.000, 0.947, -0.757, -1.532, 0.505, 1.573, 1.714, 1.233, 0.577, -0.146, -0.636, -1.696, -2.184, -2.126, -1.760, -1.623, -1.320, -1.457, -1.178, -0.948,
-    -0.659, -0.498, -0.207, 0.019, 0.081, 0.499, 1.322, 1.205, 2.155, 3.943, 0.812, -1.021, -0.258, -0.574, 0.479, 1.087, 2.013, 1.105, 0.369, -0.862,
-    -0.426, -1.630, -1.340, -1.553, -2.088, -3.090, -0.980, -2.520, -1.673, -1.000, -0.664, -0.287, 0.000, -0.104, -1.060, -1.060, 1.911, 2.198, 0.000, 0.000,
-    -0.033, -1.473, -0.321, 0.127, 0.555, 0.960, 1.789, 0.824, 0.527, -0.475, -1.896, -0.971, 0.000, 0.000, 0.000, -0.927, 0.000, 0.000, 0.000, 0.000,
-    0.000, 0.778, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.456, -0.825, -0.460, -1.518, 0.861, 0.785, 0.679, 0.679, 0.000, 0.391,
-    0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    0.000, -0.044, -0.332, -0.186, -0.060, 1.174, 1.212, 0.624, 0.438, 0.162, -0.529, -1.498, -1.921, -1.943, -1.594, -1.477, -1.627, -0.990, -0.709, -0.789,
-    -0.730, -0.351, -0.398, -0.089, 0.058, -0.327, 0.169, 1.170, 0.816, 0.000, 0.000, 0.000, -0.010, -0.047, 0.064, 0.586, 1.516, 0.925, 0.290, -0.474,
-    -0.743, -1.397, -1.435, -2.071, -2.052, -1.656, -1.040, 0.000, -1.873, -0.825, -1.451, -1.502, 0.000, -0.416, -1.279, 0.108, -0.586, 0.801, 0.000, 0.000,
-    0.862, 0.639, 0.256, -0.237, 0.051, 0.908, 1.373, 0.797, 1.150, -0.391, 0.000, 0.000, -1.389, 0.000, 0.000, 0.000, 0.000, 0.000, -0.524, 0.000,
-    0.000, 0.000, 0.000, 0.862, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.515, -0.679, -0.478, 0.000, 0.466, 1.249, 0.620, 1.314, 0.000, 0.000,
-    0.000, -0.227, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    0.000, 0.000, -1.903, -0.667, -0.033, 0.796, 0.931, 0.383, 0.170, 0.030, -0.388, -1.294, -1.481, -1.470, -1.550, -1.152, -1.192, -0.855, -0.320, -0.378,
-    -0.211, -0.241, -0.496, -0.307, -0.375, -0.306, -0.112, -0.512, 0.004, 0.623, 0.000, -0.330, -0.362, -0.599, -0.542, 0.599, 1.118, 0.577, 0.445, -0.171,
-    -0.415, -1.419, -1.229, -1.472, -0.923, -1.598, -1.141, -2.450, -0.330, -2.322, -1.103, 0.000, -0.678, -0.186, -0.330, -0.678, 0.000, 0.364, 0.000, 0.000,
-    0.000, 0.508, -0.974, -0.448, -0.473, 0.508, 1.177, 0.459, 0.636, -0.473, -0.878, -0.280, -1.466, 0.000, 0.000, -1.284, 0.000, 0.000, 0.000, 0.000,
-    0.000, 0.038, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.713, -0.992, 0.425, -0.329, -0.673, 0.020, 0.895, 1.001, 0.713, 0.307,
-    0.000, 0.000, 0.000, 1.406, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    0.000, 0.000, 0.000, -0.590, -0.198, 0.494, 0.665, 0.299, 0.043, -0.030, -0.439, -0.957, -1.391, -1.409, -1.189, -0.874, -0.991, -0.428, -0.147, 0.102,
-    -0.033, -0.239, -0.390, -0.184, -0.108, 0.010, 0.299, 0.141, -0.180, 0.593, 1.490, 0.000, -0.787, -0.675, -0.495, 0.296, 0.880, 0.530, 0.175, 0.065,
-    -0.601, -0.589, -1.645, -1.112, -1.100, -1.372, -1.554, 0.430, -0.937, -0.994, -0.669, 0.000, -0.381, -0.907, -0.119, 0.479, -0.907, 2.184, 0.000, 0.000,
-    0.000, 0.057, -1.121, -0.754, 0.057, -0.166, 1.213, 0.345, 0.394, -0.954, -0.954, 0.000, -0.428, -0.428, -0.754, -0.261, 0.000, 0.000, 0.000, 0.000,
-    0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.796, -0.708, 0.023, 0.642, 0.000, 0.000, 1.202, 0.796, 0.000, -0.120,
-    -0.302, 0.000, 0.000, 0.000, 0.391, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    0.000, -1.576, -1.587, -1.047, -0.641, 0.065, 0.263, -0.051, -0.159, -0.313, -0.366, -0.428, -0.566, -0.484, -0.531, -0.401, -0.277, 0.021, 0.388, 0.496,
-    0.289, 0.271, 0.452, 0.289, 0.436, 0.609, 0.633, 0.590, 0.835, 1.196, -0.506, -2.100, -0.606, -0.496, -1.190, -0.179, 0.562, 0.324, 0.105, -0.130,
-    -0.185, -0.332, -0.011, -0.262, -0.407, -0.475, -0.125, -0.244, 0.059, 0.358, 0.066, 0.094, -0.158, 0.278, -0.339, 0.066, 1.129, 0.000, 0.000, 0.000,
-    -1.370, -0.633, -0.979, -1.273, -0.531, 0.040, 0.487, 0.954, 0.088, 0.170, -0.254, 0.124, 0.366, -0.307, -1.034, 0.576, 0.576, -0.871, 0.170, -1.439,
-    -1.565, -0.235, 0.000, 0.000, 0.000, -0.929, 0.576, 0.000, 0.000, 0.000, -0.531, -0.810, -0.029, -0.413, -0.354, 0.098, 0.931, 0.871, 0.417, 0.024,
-    -1.586, 0.280, -0.268, -0.651, 0.242, 1.053, 0.648, 1.746, 0.000, 0.648, 1.053, 1.053, 0.000, 1.053, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000,
-    -0.793, -0.520, -0.655, -0.663, -0.588, -0.522, -0.520, -0.537, -0.474, -0.368, -0.244, -0.053, 0.073, 0.032, 0.109, 0.085, 0.155, 0.331, 0.458, 0.538,
-    0.552, 0.593, 0.776, 0.884, 1.057, 1.289, 1.542, 2.035, 2.786, 3.059, 0.262, -0.615, -0.559, -0.597, -0.572, -0.139, 0.111, 0.075, -0.024, -0.026,
-    0.100, 0.151, 0.351, 0.410, 0.412, 0.470, 0.588, 0.694, 0.619, 0.729, 0.668, 0.741, 1.049, 0.982, 0.611, 1.213, 1.458, 0.899, 0.000, 0.000,
-    0.104, 0.171, -0.685, -0.394, -0.431, -0.125, 0.419, -0.066, -0.111, 0.387, 0.374, 0.358, 0.374, 0.676, 0.397, 0.586, 0.770, 1.441, 1.052, 0.802,
-    0.958, 1.564, 1.672, 1.208, 2.093, 2.365, 2.722, 0.000, 0.000, 0.000, -0.311, -0.291, -0.540, -0.113, 0.188, 0.466, 0.774, 0.390, 0.337, 0.434,
-    1.088, 0.227, 1.188, 1.658, 0.434, 0.098, 0.029, 0.000, -0.076, 0.434, 0.434, 1.128, 0.000, 0.840, -0.952, 0.000, 0.000, 0.000, 0.000, 0.000,
-    };
-
-
-  };
-}
-
-#endif
